requirements-actions.txt
pytablewriter
requests
ruamel.yaml
glpi==0.4.0
future